refactor: util 패키지와 상수 클래스 따로 분리
