+++ src/main/java/baseball/util/BaseballConst.java
@@ -1,0 +1,8 @@
+package baseball.util;
+
+public class BaseballConst {
+
+    public static final int MAX_BALL_COUNT = 3;
+
+    public static final int STRIKE_OUT = 3;
+}
