+++ src/main/java/baseball/util/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package baseball.core;
+package baseball.util;
 
 import camp.nextstep.edu.missionutils.Randoms;
 
@@ -6,11 +6,12 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import static baseball.util.BaseballConst.MAX_BALL_COUNT;
+
 public class RandomNumberGenerator {
 
     private static final int START_INCLUSIVE = 1;
     private static final int END_INCLUSIVE = 9;
-    private static final int MAX_BALL_COUNT = 3;
 
     public static List<Integer> generate() {
         return Stream.generate(() -> Randoms.pickNumberInRange(START_INCLUSIVE, END_INCLUSIVE))
